feat: promote shared_validation-based v2 validators to production, archive originals as legacy
Encounters and discovery validators are now built on shared_validation/,
invoked directly by encounters_master_validator.py and
discovery_master_validator.py (no wrapper changes needed — both call by
filename, and the promoted files now occupy those filenames).

- validate_encounters.py / validate_discovery.py: promoted from _v2,
  docstrings rewritten to describe current (live) status instead of
  "comparison only."
- legacy/validate_encounters_legacy.py, legacy/validate_discovery_legacy.py:
  pre-migration originals archived intact, not deleted. Each carries a
  loud ARCHIVED header and a legacy/README.md marking it as dead code.
- legacy/verify_image_urls.py: a real, standalone copy for encounters'
  legacy validator, so the archive never depends on (or risks) the live
  copy. Intentional drift over time, not a bug.
- Bug found and fixed during the move: both legacy scripts hardcoded a
  directory-depth assumption (SCRIPTS_DIR.parent) that broke once moved
  one level deeper into legacy/ — fixed, and both legacy scripts verified
  byte-identical in output to the pre-migration baseline.
- tests/test_promoted_validators.py: durable smoke-test gate — shells out
  to the real master validators against real content (clean-pass case),
  plus isolated temp fixtures with deliberately broken index.json
  (gate-on-failure case), so this migration's correctness is CI-enforced
  going forward rather than a one-time manual verification claim.
- shared_validation/README.md updated to reflect its new load-bearing,
  production status.

Design and migration plan independently reviewed by two separate architect
passes (SOLID principles + safe-migration practice) before implementation.
